+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 - Chg: Remove `'sort' => 'older'` for the "Spaces" top menu entry
 - Enh: Added `transition/embed/mail-conversation-sidebar`
 - Enh: Added `transition/embed/notification-overview`
+- Cgh #1041: Changed @background-color-page to #f0f2f5'
 
 1.7.1 (Nov 27, 2023)
 --------------------